Split common parts of executions monitoring into executions monitoring component.
